+++ db/migrate/20131017064039_repair_work_packages_initial_attachable_journal.rb
@@ -13,6 +13,7 @@ require_relative 'migration_utils/attachable_utils'
 
 class RepairWorkPackagesInitialAttachableJournal < ActiveRecord::Migration
   include Migration::Utils
+  include Migration::Utils::AttachableUtils
 
   LEGACY_JOURNAL_TYPE = 'IssueJournal'
   JOURNAL_TYPE = 'WorkPackage'